requirements.txt
flask
line-bot-sdk
openai
python-dotenv
requests
beautifulsoup4
gunicorn
